Add validation utilities and UI components for register authentication
- Implemented AppRegex class for email, password, and phone number validation.
- Created AppAssets class for managing asset paths.
- Developed spacing widgets for consistent layout.
- Built Fields widget for user input with validation for first name, last name, phone number, email, password, confirm password, and city ID.
- Added PickAvatar widget for selecting user avatar from gallery.

+++ lib/core/widgets/app_text_button.dart
@@ -14,8 +14,8 @@ class AppTextButton extends StatelessWidget {
   final String text;
   final bool isLoading;
   final TextStyle? textStyle;
-  final int? width;
-  final int? height;
+  final double? width;
+  final double? height;
 
   @override
   Widget build(BuildContext context) {
